temp/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-mail>=0.10.0",
    "mysql-connector-python>=9.3.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "werkzeug>=3.1.3",
]
